# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend Challenge",
        "version": "1.0.0",
        "title": "Turing/E-commerce",
        "contact": {
            "email": "codeninja0x01@gmail.com"
        }
    },
    "basePath": "/api",
...
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "bearer": {
            "type": "apiKey",
            "name": "USER-KEY",
            "in": "header"
        }
    },
    "host": "localhost:3000",
    "paths": {
...
        "/categories": {
            "get": {
                "summary": "Get categories.",
                "description": "Return list of categories.",
                "parameters": [
                    {
                        "$ref": "#/parameters/limitParam"
                    },
                    {
                        "$ref": "#/parameters/offsetParam"
                    },
                    {
                        "$ref": "#/parameters/orderParam"
                    },
                    {
                        "$ref": "#/parameters/orderByParam"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "$ref": "#/definitions/CategoryPagination"
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "category"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/categories/{id}": {
            "get": {
                "summary": "Get category.",
                "description": "Return category by Id.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "id",
                        "required": true,
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "$ref": "#/definitions/Category"
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "category"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/categories/product/{productId}": {
            "get": {
                "summary": "Get category of product.",
                "description": "Return category by productId.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "productId",
                        "required": true,
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Category"
                            }
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "category"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/categories/department/{departmentId}": {
            "get": {
                "summary": "Get category of department.",
                "description": "Return category by departmentId.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "departmentId",
                        "required": true,
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Category"
                            }
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "category"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
...
        "/orders": {
            "post": {
                "summary": "Add order.",
                "description": "Create order.",
                "parameters": [
                    {
                        "name": "cartKey",
                        "required": true,
                        "in": "formData",
                        "type": "string"
                    },
                    {
                        "name": "taxId",
                        "required": true,
                        "in": "formData",
                        "type": "number"
                    },
                    {
                        "name": "shippingId",
                        "required": true,
                        "in": "formData",
                        "type": "number"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successfully operation.",
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    },
                    "401": {
                        "description": "Unauthorized User."
                    }
                },
                "tags": [
                    "order"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ]
            }
        },
        "/orders/customer": {
            "get": {
                "summary": "Get user order.",
                "description": "Return user list of order.",
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Order"
                            }
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    },
                    "401": {
                        "description": "Unauthorized User."
                    }
                },
                "tags": [
                    "order"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/orders/{id}": {
            "get": {
                "summary": "Get order.",
                "description": "Return order by id.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "id",
                        "required": true,
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Order"
                            }
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    },
                    "401": {
                        "description": "Unauthorized User."
                    }
                },
                "tags": [
                    "order"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/orders/{id}/detail": {
            "get": {
                "summary": "Get order detail.",
                "description": "Return order detail by orderId.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "id",
                        "required": true,
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/OrderDetail"
                            }
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    },
                    "401": {
                        "description": "Unauthorized User."
                    }
                },
                "tags": [
                    "order"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
...
        "/products/category/{categoryId}": {
            "get": {
                "summary": "Get category products.",
                "description": "Return products by categoryId.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "categoryId",
                        "required": true,
                        "in": "path"
                    },
                    {
                        "$ref": "#/parameters/limitParam"
                    },
                    {
                        "$ref": "#/parameters/offsetParam"
                    },
                    {
                        "$ref": "#/parameters/orderParam"
                    },
                    {
                        "$ref": "#/parameters/orderByParam"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Product"
                            }
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "product"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
...
    },
    "parameters": {
        "limitParam": {
            "name": "limit",
            "in": "query",
            "description": "Limits the number of returned results",
            "required": false,
            "type": "integer",
            "default": 20
        },
        "offsetParam": {
            "name": "offset",
            "in": "query",
            "description": "Offset from which start returned results",
            "required": false,
            "type": "integer",
            "default": 1
        },
        "orderParam": {
            "name": "order",
            "in": "query",
            "description": "Order ASC or Desc the results",
            "required": false,
            "type": "string",
            "enum": [
                "ASC",
                "DESC"
            ]
        },
        "orderByParam": {
            "name": "orderBy",
            "in": "query",
            "description": "OrderBy name or id the results",
            "required": false,
            "type": "string",
            "enum": [
                "id",
                "name"
            ]
        },
...
    },
    "definitions": {
...
        "Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            },
            "required": [
                "id",
                "name",
                "description"
            ]
        },
        "CategoryPagination": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "number"
                },
                "currentPage": {
                    "type": "number"
                },
                "perPage": {
                    "type": "number"
                },
                "row": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Category"
                    }
                }
            },
            "required": [
                "total",
                "currentPage",
                "perPage",
                "row"
            ]
        },
...
        "Order": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "number"
                },
                "totalAmount": {
                    "type": "number"
                },
                "shippedOn": {
                    "type": "string",
                    "format": "date-time"
                },
                "status": {
                    "type": "number"
                },
                "comments": {
                    "type": "string"
                },
                "authCode": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string",
                    "format": "date-time"
                }
            },
            "required": [
                "id",
                "totalAmount",
                "shippedOn",
                "status",
                "comments",
                "authCode",
                "reference",
                "createdAt"
            ]
        },
        "OrderDetail": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "number"
                },
                "attribute": {
                    "type": "string"
                },
                "productName": {
                    "type": "string"
                },
                "quantity": {
                    "type": "number"
                },
                "unitCost": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                }
            },
            "required": [
                "id",
                "attribute",
                "productName",
                "quantity",
                "unitCost",
                "currency"
            ]
        },
...
        "Product": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "discountedPrice": {
                    "type": "number"
                },
                "image1": {
                    "type": "string"
                },
                "image2": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                },
                "display": {
                    "type": "boolean"
                }
            },
            "required": [
                "id",
                "name",
                "description",
                "price",
                "discountedPrice",
                "image1",
                "image2",
                "thumbnail",
                "display"
            ]
        },
...
    }
}